Edited tables and types for columns

+++ express-back-end/db/migrations/20190622065301_create_moods.js
@@ -5,8 +5,8 @@ exports.up = function(knex, Promise) {
       table.increments('id');
       table.integer('user_id').references('users.id')
       table.date('date');
-      table.integer('rank');
-      table.integer('emoji_rank');
+      table.float('rank');
+      table.float('emoji_rank');
     })
   ])
 };
