+++ express-back-end/db/migrations/20190622067646_create_responses.js
@@ -6,7 +6,7 @@ exports.up = function(knex, Promise) {
       table.integer('mood_id').references('moods.id')
       table.integer('question_id').references('questions.id')
       table.string('answer');
-      table.integer('sentiment_rank');
+      table.float('sentiment_rank');
     })
   ])
 };
